requirements.txt
flask
python-telegram-bot==20.4
SQLAlchemy
psycopg2-binary
gunicorn
